Primary colour change and context indicator
Previous primary colour was not accessible. Have change the lighter teal colour to a darker cyan colour.
Have also introduced an element in the global header that that tells the user which dataset they are browsing in. Indicated by a 'pencil' icon, clicking the name of the dataset will present a modal to the user communicating the name and type of the dataset and offer the ability to change the dataset.